refactor: mirror unit test folders by responsibility
