+++ tests/SemanticStub.Api.Tests/Unit/Controllers/StubControllerTests.cs
@@ -9,7 +9,7 @@ using SemanticStub.Api.Models;
 using SemanticStub.Api.Services;
 using Xunit;
 
-namespace SemanticStub.Api.Tests.Unit;
+namespace SemanticStub.Api.Tests.Unit.Controllers;
 
 public sealed class StubControllerTests
 {
