+++ tests/SemanticStub.Api.Tests/Unit/Inspection/StubInspectionProjectionBuilderTests.cs
@@ -4,7 +4,7 @@ using SemanticStub.Api.Models;
 using SemanticStub.Api.Services;
 using Xunit;
 
-namespace SemanticStub.Api.Tests.Unit;
+namespace SemanticStub.Api.Tests.Unit.Inspection;
 
 public sealed class StubInspectionProjectionBuilderTests
 {
