+++ tests/SemanticStub.Api.Tests/Unit/Resolution/StubResponseBuilderTests.cs
@@ -2,7 +2,7 @@ using SemanticStub.Api.Models;
 using SemanticStub.Api.Services;
 using Xunit;
 
-namespace SemanticStub.Api.Tests.Unit;
+namespace SemanticStub.Api.Tests.Unit.Resolution;
 
 public sealed class StubResponseBuilderTests
 {
