+++ tests/SemanticStub.Api.Tests/Unit/Resolution/StubServiceTests.cs
@@ -7,7 +7,7 @@ using SemanticStub.Api.Models;
 using SemanticStub.Api.Services;
 using Xunit;
 
-namespace SemanticStub.Api.Tests.Unit;
+namespace SemanticStub.Api.Tests.Unit.Resolution;
 
 public sealed class StubServiceTests
 {
